# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Equipment Tracker API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Equipment {
  id: number;
  tileUuid: string;
  qrToken: string;
  /** @nullable */
  customQrCode?: string | null;
  label: string;
  category: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  serialNumber?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  inServiceDate?: string | null;
  /** @nullable */
  outOfServiceDate?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface TileDevice {
  uuid: string;
  name: string;
  kind: string;
  archetype: string;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  /** @nullable */
  altitude?: number | null;
  /** @nullable */
  accuracy?: number | null;
  /** @nullable */
  lastSeen?: string | null;
  lost: boolean;
  dead: boolean;
  firmwareVersion: string;
  hardwareVersion: string;
  equipment?: Equipment | null;
}

export interface LocationPoint {
  latitude: number;
  longitude: number;
  /** @nullable */
  altitude?: number | null;
  /** @nullable */
  accuracy?: number | null;
  timestamp: string;
}

export interface EquipmentInput {
  tileUuid: string;
  customQrCode?: string;
  label: string;
  category: string;
  description?: string;
  serialNumber?: string;
  notes?: string;
  inServiceDate?: string;
  outOfServiceDate?: string;
}

export interface EquipmentUpdate {
  customQrCode?: string;
  label?: string;
  category?: string;
  description?: string;
  serialNumber?: string;
  notes?: string;
  inServiceDate?: string;
  outOfServiceDate?: string;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface DashboardSummary {
  total: number;
  active: number;
  lost: number;
  dead: number;
  byCategory: CategoryCount[];
  recentlyLost: TileDevice[];
}

export interface QrScan {
  id: number;
  equipmentId: number;
  latitude: number;
  longitude: number;
  /** @nullable */
  accuracy?: number | null;
  /** @nullable */
  city?: string | null;
  /** @nullable */
  userAgent?: string | null;
  scannedAt: string;
}

export interface ScanInfo {
  equipmentId: number;
  label: string;
  category: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  serialNumber?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  inServiceDate?: string | null;
  /** @nullable */
  outOfServiceDate?: string | null;
  lastQrScan?: QrScan | null;
  totalScans: number;
}

export interface ScanLocationInput {
  latitude: number;
  longitude: number;
  accuracy?: number;
}

export interface EquipmentLog {
  id: number;
  equipmentId: number;
  logType: string;
  logDate: string;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  operatorName?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface EquipmentLogInput {
  logType: string;
  logDate: string;
  durationMinutes?: number;
  operatorName?: string;
  location?: string;
  notes?: string;
}

export interface QrLookupResult {
  equipment: Equipment;
  tile?: TileDevice | null;
  lastQrScan?: QrScan | null;
  recentLogs: EquipmentLog[];
}

export type QrLookupParams = {
code: string;
};