# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Equipment Tracker API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: tiles
  - name: equipment
  - name: history
  - name: scan
  - name: logs
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /tiles:
    get:
      operationId: getTiles
      tags: [tiles]
      summary: Get all Tiles
      responses:
        "200":
          description: List of tiles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TileDevice"
        "401":
          description: Tile auth failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /tiles/{tileUuid}/history:
    get:
      operationId: tileHistory
      tags: [history]
      summary: Get location history for a Tile
      parameters:
        - name: tileUuid
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Location history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LocationPoint"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /equipment:
    get:
      operationId: listEquipment
      tags: [equipment]
      summary: List all equipment
      responses:
        "200":
          description: List of equipment
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Equipment"
    post:
      operationId: createEquipment
      tags: [equipment]
      summary: Create or update equipment for a Tile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipmentInput"
      responses:
        "201":
          description: Equipment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"

  /equipment/{id}:
    get:
      operationId: getEquipment
      tags: [equipment]
      summary: Get a single equipment record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Equipment record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateEquipment
      tags: [equipment]
      summary: Update equipment metadata
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipmentUpdate"
      responses:
        "200":
          description: Updated equipment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Equipment"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteEquipment
      tags: [equipment]
      summary: Delete equipment record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /equipment/{id}/scans:
    get:
      operationId: listQrScans
      tags: [scan]
      summary: Get QR scan history for equipment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scan history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QrScan"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /equipment/{id}/logs:
    get:
      operationId: listEquipmentLogs
      tags: [logs]
      summary: Get use logs for equipment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Use logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EquipmentLog"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createEquipmentLog
      tags: [logs]
      summary: Add a use log entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EquipmentLogInput"
      responses:
        "201":
          description: Log created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EquipmentLog"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /equipment/{id}/logs/{logId}:
    delete:
      operationId: deleteEquipmentLog
      tags: [logs]
      summary: Delete a use log entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: logId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [equipment]
      summary: Get dashboard summary stats
      responses:
        "200":
          description: Summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"

  /scan/{token}:
    get:
      operationId: getScanInfo
      tags: [scan]
      summary: Get equipment info by QR token (public)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Equipment info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanInfo"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: recordScan
      tags: [scan]
      summary: Record a QR scan with GPS location
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanLocationInput"
      responses:
        "200":
          description: Scan recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrScan"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /qr/lookup:
    get:
      operationId: qrLookup
      tags: [scan]
      summary: Look up equipment by any QR code value
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Equipment found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QrLookupResult"
        "404":
          description: No equipment matches this code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    TileDevice:
      type: object
      properties:
        uuid:
          type: string
        name:
          type: string
        kind:
          type: string
        archetype:
          type: string
        latitude:
          type: ["number", "null"]
        longitude:
          type: ["number", "null"]
        altitude:
          type: ["number", "null"]
        accuracy:
          type: ["number", "null"]
        lastSeen:
          type: ["string", "null"]
        lost:
          type: boolean
        dead:
          type: boolean
        firmwareVersion:
          type: string
        hardwareVersion:
          type: string
        equipment:
          $ref: "#/components/schemas/Equipment"
          nullable: true
      required: [uuid, name, kind, archetype, lost, dead, firmwareVersion, hardwareVersion]

    LocationPoint:
      type: object
      properties:
        latitude:
          type: number
        longitude:
          type: number
        altitude:
          type: ["number", "null"]
        accuracy:
          type: ["number", "null"]
        timestamp:
          type: string
      required: [latitude, longitude, timestamp]

    Equipment:
      type: object
      properties:
        id:
          type: integer
        tileUuid:
          type: string
        qrToken:
          type: string
        customQrCode:
          type: ["string", "null"]
        label:
          type: string
        category:
          type: string
        description:
          type: ["string", "null"]
        serialNumber:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        inServiceDate:
          type: ["string", "null"]
        outOfServiceDate:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, tileUuid, qrToken, label, category, createdAt, updatedAt]

    EquipmentInput:
      type: object
      properties:
        tileUuid:
          type: string
        customQrCode:
          type: string
        label:
          type: string
        category:
          type: string
        description:
          type: string
        serialNumber:
          type: string
        notes:
          type: string
        inServiceDate:
          type: string
        outOfServiceDate:
          type: string
      required: [tileUuid, label, category]

    EquipmentUpdate:
      type: object
      properties:
        customQrCode:
          type: string
        label:
          type: string
        category:
          type: string
        description:
          type: string
        serialNumber:
          type: string
        notes:
          type: string
        inServiceDate:
          type: string
        outOfServiceDate:
          type: string

    DashboardSummary:
      type: object
      properties:
        total:
          type: integer
        active:
          type: integer
        lost:
          type: integer
        dead:
          type: integer
        byCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategoryCount"
        recentlyLost:
          type: array
          items:
            $ref: "#/components/schemas/TileDevice"
      required: [total, active, lost, dead, byCategory, recentlyLost]

    CategoryCount:
      type: object
      properties:
        category:
          type: string
        count:
          type: integer
      required: [category, count]

    QrScan:
      type: object
      properties:
        id:
          type: integer
        equipmentId:
          type: integer
        latitude:
          type: number
        longitude:
          type: number
        accuracy:
          type: ["number", "null"]
        city:
          type: ["string", "null"]
        userAgent:
          type: ["string", "null"]
        scannedAt:
          type: string
      required: [id, equipmentId, latitude, longitude, scannedAt]

    ScanInfo:
      type: object
      properties:
        equipmentId:
          type: integer
        label:
          type: string
        category:
          type: string
        description:
          type: ["string", "null"]
        serialNumber:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        inServiceDate:
          type: ["string", "null"]
        outOfServiceDate:
          type: ["string", "null"]
        lastQrScan:
          $ref: "#/components/schemas/QrScan"
          nullable: true
        totalScans:
          type: integer
      required: [equipmentId, label, category, totalScans]

    ScanLocationInput:
      type: object
      properties:
        latitude:
          type: number
        longitude:
          type: number
        accuracy:
          type: number
      required: [latitude, longitude]

    EquipmentLog:
      type: object
      properties:
        id:
          type: integer
        equipmentId:
          type: integer
        logType:
          type: string
        logDate:
          type: string
        durationMinutes:
          type: ["integer", "null"]
        operatorName:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, equipmentId, logType, logDate, createdAt]

    EquipmentLogInput:
      type: object
      properties:
        logType:
          type: string
        logDate:
          type: string
        durationMinutes:
          type: integer
        operatorName:
          type: string
        location:
          type: string
        notes:
          type: string
      required: [logType, logDate]

    QrLookupResult:
      type: object
      properties:
        equipment:
          $ref: "#/components/schemas/Equipment"
        tile:
          $ref: "#/components/schemas/TileDevice"
          nullable: true
        lastQrScan:
          $ref: "#/components/schemas/QrScan"
          nullable: true
        recentLogs:
          type: array
          items:
            $ref: "#/components/schemas/EquipmentLog"
      required: [equipment, recentLogs]