# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Equipment Tracker API
 * OpenAPI spec version: 0.1.0
 */
import type { CategoryCount } from './categoryCount';
import type { TileDevice } from './tileDevice';

export interface DashboardSummary {
  total: number;
  active: number;
  lost: number;
  dead: number;
  byCategory: CategoryCount[];
  recentlyLost: TileDevice[];
}