# lib/api-zod/src/generated/types/equipment.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Equipment Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface Equipment {
  id: number;
  tileUuid: string;
  qrToken: string;
  /** @nullable */
  customQrCode?: string | null;
  label: string;
  category: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  serialNumber?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  inServiceDate?: string | null;
  /** @nullable */
  outOfServiceDate?: string | null;
  createdAt: string;
  updatedAt: string;
}