# lib/api-zod/src/generated/types/equipmentLog.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Equipment Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface EquipmentLog {
  id: number;
  equipmentId: number;
  logType: string;
  logDate: string;
  /** @nullable */
  durationMinutes?: number | null;
  /** @nullable */
  operatorName?: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}