# lib/api-zod/src/generated/types/equipmentLogInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Equipment Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface EquipmentLogInput {
  logType: string;
  logDate: string;
  durationMinutes?: number;
  operatorName?: string;
  location?: string;
  notes?: string;
}