# lib/api-zod/src/generated/types/equipmentUpdate.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Equipment Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface EquipmentUpdate {
  customQrCode?: string;
  label?: string;
  category?: string;
  description?: string;
  serialNumber?: string;
  notes?: string;
  inServiceDate?: string;
  outOfServiceDate?: string;
}