# lib/api-zod/src/generated/types/locationPoint.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Equipment Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface LocationPoint {
  latitude: number;
  longitude: number;
  /** @nullable */
  altitude?: number | null;
  /** @nullable */
  accuracy?: number | null;
  timestamp: string;
}