# lib/api-zod/src/generated/types/scanLocationInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Equipment Tracker API
 * OpenAPI spec version: 0.1.0
 */

export interface ScanLocationInput {
  latitude: number;
  longitude: number;
  accuracy?: number;
}